refactor: sequential subproject
